test: Jpa test와 기존 테스트 분리
