+++ src/test/java/roomescape/JpaTest.java
@@ -1,0 +1,31 @@
+package roomescape;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
+import roomescape.time.Time;
+import roomescape.time.TimeRepository;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+@DataJpaTest
+public class JpaTest {
+    @Autowired
+    private TestEntityManager entityManager;
+
+    @Autowired
+    private TimeRepository timeRepository;
+
+    @Test
+    void 사단계() {
+        Time time = new Time("10:00");
+        entityManager.persist(time);
+        entityManager.flush();
+
+        Time persistTime = timeRepository.findById(time.getId()).orElse(null);
+
+        assertThat(persistTime.getTime()).isEqualTo(time.getTime());
+    }
+}
+
