+++ src/test/java/roomescape/MissionStepTest.java
@@ -5,14 +5,9 @@ import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import roomescape.reservation.ReservationResponse;
-import roomescape.time.Time;
-import roomescape.time.TimeRepository;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -109,25 +104,5 @@ public class MissionStepTest {
                 .get("/admin")
                 .then().log().all()
                 .statusCode(200);
-    }
-
-    @DataJpaTest
-    public class JpaTest {
-        @Autowired
-        private TestEntityManager entityManager;
-
-        @Autowired
-        private TimeRepository timeRepository;
-
-        @Test
-        void 사단계() {
-            Time time = new Time("10:00");
-            entityManager.persist(time);
-            entityManager.flush();
-
-            Time persistTime = timeRepository.findById(time.getId()).orElse(null);
-
-            assertThat(persistTime.getTime()).isEqualTo(time.getTime());
-        }
     }
 }
